refac: graphql-codegen.ts
